WIP refactored into appropriate directories
